refactor : HttpClient 리턴 타입 변경
