+++ src/common/network/AxiosHttpClient.ts
@@ -18,7 +18,7 @@ export default class AxiosHttpClient implements IHttpClient {
     });
   }
 
-  get<T>(endPoint: string, options?: any): Promise<ApiResponse<T>> {
+  get<T>(endPoint: string, options?: any): Promise<SuccessResponse<T>> {
     return this.client
       .get(endPoint, options)
       .then(
@@ -41,7 +41,7 @@ export default class AxiosHttpClient implements IHttpClient {
     endPoint: string,
     payload: any,
     options?: any,
-  ): Promise<ApiResponse<T>> {
+  ): Promise<SuccessResponse<T>> {
     return this.client
       .post(endPoint, payload, options)
       .then(
@@ -61,7 +61,7 @@ export default class AxiosHttpClient implements IHttpClient {
     endPoint: string,
     payload: any,
     options?: any,
-  ): Promise<ApiResponse<T>> {
+  ): Promise<SuccessResponse<T>> {
     return this.client
       .put(endPoint, payload, options)
       .then(
@@ -77,7 +77,7 @@ export default class AxiosHttpClient implements IHttpClient {
       });
   }
 
-  delete(endPoint: string, options?: any): Promise<ApiResponse<boolean>> {
+  delete(endPoint: string, options?: any): Promise<SuccessResponse<boolean>> {
     return this.client
       .delete(endPoint, options)
       .then(
