+++ src/common/network/types/HttpClient.interface.ts
@@ -1,20 +1,20 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import { ApiResponse } from './apiResponse';
+import { ApiResponse, SuccessResponse } from './apiResponse';
 
 export interface IHttpClient {
-  get<T>(endPoint: string, options?: any): Promise<ApiResponse<T>>;
+  get<T>(endPoint: string, options?: any): Promise<SuccessResponse<T>>;
 
   post<T>(
     endPoint: string,
     payload: any,
     options?: any,
-  ): Promise<ApiResponse<T>>;
+  ): Promise<SuccessResponse<T>>;
 
   put<T>(
     endPoint: string,
     payload: any,
     options?: any,
-  ): Promise<ApiResponse<T>>;
+  ): Promise<SuccessResponse<T>>;
 
-  delete(endPoint: string, options?: any): Promise<ApiResponse<boolean>>;
+  delete(endPoint: string, options?: any): Promise<SuccessResponse<boolean>>;
 }
